Put test helper stuff in test_helper.exs
