--- test/support/temp_utils.ex
@@ -1,15 +1,0 @@
-defmodule Support.TempUtils do
-  @moduledoc """
-  Utilities for dealing with temporary files.
-  """
-
-  @doc """
-  Generates a uniquely-named file in the system temp directory.
-  """
-  def temp_path(ext \\ ".md") do
-    dir = System.tmp_dir!()
-    filename = "#{System.unique_integer([:positive])}#{ext}"
-
-    Path.join(dir, filename)
-  end
-end
